+++ lib/label/label.c
@@ -1871,15 +1871,15 @@ bool dev_read_bytes(struct device *dev, uint64_t start, size_t len, void *data)
 	if (dev->bcache_di < 0) {
 		/* This is not often needed. */
 		if (!label_scan_open(dev)) {
-			log_error("Error opening device %s for reading at %llu length %u.",
-				  dev_name(dev), (unsigned long long)start, (uint32_t)len);
+			log_error("Error opening device %s for reading at %llu length %zu.",
+				  dev_name(dev), (unsigned long long)start, len);
 			return false;
 		}
 	}
 
 	if (!bcache_read_bytes(scan_bcache, dev->bcache_di, start, len, data)) {
-		log_error("Error reading device %s at %llu length %u.",
-			  dev_name(dev), (unsigned long long)start, (uint32_t)len);
+		log_error("Error reading device %s at %llu length %zu.",
+			  dev_name(dev), (unsigned long long)start, len);
 		label_scan_invalidate(dev);
 		return false;
 	}
@@ -1912,23 +1912,23 @@ bool dev_write_bytes(struct device *dev, uint64_t start, size_t len, void *data)
 		/* This is not often needed. */
 		dev->flags |= DEV_BCACHE_WRITE;
 		if (!label_scan_open(dev)) {
-			log_error("Error opening device %s for writing at %llu length %u.",
-				  dev_name(dev), (unsigned long long)start, (uint32_t)len);
+			log_error("Error opening device %s for writing at %llu length %zu.",
+				  dev_name(dev), (unsigned long long)start, len);
 			return false;
 		}
 	}
 
 	if (!bcache_write_bytes(scan_bcache, dev->bcache_di, start, len, data)) {
-		log_error("Error writing device %s at %llu length %u.",
-			  dev_name(dev), (unsigned long long)start, (uint32_t)len);
+		log_error("Error writing device %s at %llu length %zu.",
+			  dev_name(dev), (unsigned long long)start, len);
 		dev_unset_last_byte(dev);
 		label_scan_invalidate(dev);
 		return false;
 	}
 
 	if (!bcache_flush(scan_bcache)) {
-		log_error("Error writing device %s at %llu length %u.",
-			  dev_name(dev), (unsigned long long)start, (uint32_t)len);
+		log_error("Error writing device %s at %llu length %zu.",
+			  dev_name(dev), (unsigned long long)start, len);
 		dev_unset_last_byte(dev);
 		label_scan_invalidate(dev);
 		return false;
@@ -1978,8 +1978,8 @@ bool dev_set_bytes(struct device *dev, uint64_t start, size_t len, uint8_t val)
 		/* This is not often needed. */
 		dev->flags |= DEV_BCACHE_WRITE;
 		if (!label_scan_open(dev)) {
-			log_error("Error opening device %s for writing at %llu length %u.",
-				  dev_name(dev), (unsigned long long)start, (uint32_t)len);
+			log_error("Error opening device %s for writing at %llu length %zu.",
+				  dev_name(dev), (unsigned long long)start, len);
 			return false;
 		}
 	}
@@ -1992,14 +1992,14 @@ bool dev_set_bytes(struct device *dev, uint64_t start, size_t len, uint8_t val)
 		rv = bcache_set_bytes(scan_bcache, dev->bcache_di, start, len, val);
 
 	if (!rv) {
-		log_error("Error writing device value %s at %llu length %u.",
-			  dev_name(dev), (unsigned long long)start, (uint32_t)len);
+		log_error("Error writing device value %s at %llu length %zu.",
+			  dev_name(dev), (unsigned long long)start, len);
 		goto fail;
 	}
 
 	if (!bcache_flush(scan_bcache)) {
-		log_error("Error writing device %s at %llu length %u.",
-			  dev_name(dev), (unsigned long long)start, (uint32_t)len);
+		log_error("Error writing device %s at %llu length %zu.",
+			  dev_name(dev), (unsigned long long)start, len);
 		goto fail;
 	}
